requirements.txt
tensorflow
numpy
pillow 